feat: implement api of login

+++ src/main/java/com/br/retrobackend/entitys/Action.java
@@ -22,7 +22,7 @@ public class Action implements Serializable {
     @Setter(AccessLevel.NONE)
     @GeneratedValue(strategy = GenerationType.SEQUENCE)
     @Id
-    private int actionId;
+    private Long actionId;
 
     @NotNull
     @Column(name = "action_description")
