+++ src/main/java/com/br/retrobackend/repository/ActionRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ActionRepository extends JpaRepository<Action, Integer> { }
+public interface ActionRepository extends JpaRepository<Action, Long> { }
