+++ src/main/java/com/br/retrobackend/repository/RetrospectiveRepository.java
@@ -5,4 +5,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RetrospectiveRepository extends JpaRepository<Retrospective, Integer> { }
+public interface RetrospectiveRepository extends JpaRepository<Retrospective, Long> { }
